feat: Add mobile device support and improve layout
This commit includes changes in the layout files to add support for mobile devices.
The changes include addition of 'hideForMobile' and 'showForMobile' divs and a script
to handle the display for different screen sizes.

The calendar view has been updated to improve the display of contiguous stays
information, including the number of days in between stays.

The dark mode option in the tailwind configuration has also been set to 'false'.