feat: Add Tor and Source-Code links to navigation and login/register views
In the navigation view, the previously commented out section is now active with
two new links. The first link is to the source code on GitHub and the second
is a Tor link.

In the login view, a check is added to redirect Tor users to the onion site.
Also, an Onion/Tor link is added.

In the register view, an Onion/Tor link is also added.