feat: Update return value and improve user login functionality #123
This commit updates the return value in the instance.js file to 'signer'
instead of 'instance'. Also made changes in nostrApp.js to include the
signing of users and calling the 'loginNostr' function. Finally, in
login.blade.php, the '$login' function was renamed to '$loginNostr' and
the redirecting logic was updated.